enhan(selectors): Enhacement of the selector function

+++ utils/fct/gw_fct_setselectors.sql
@@ -330,9 +330,6 @@ BEGIN
 
 			IF v_tabname = 'tab_exploitation' THEN
 				DELETE FROM selector_macroexpl WHERE cur_user = current_user;
-				INSERT INTO selector_macroexpl
-				SELECT DISTINCT macroexpl_id, current_user FROM exploitation WHERE active is true and expl_id IN (SELECT expl_id FROM selector_expl WHERE cur_user = current_user)
-				ON CONFLICT (macroexpl_id, cur_user) DO NOTHING;
 
 			ELSIF v_tabname = 'tab_macroexploitation' THEN
 				DELETE FROM selector_expl WHERE cur_user = current_user;
@@ -343,13 +340,10 @@ BEGIN
 
 			-- macrosector
 			DELETE FROM selector_macrosector WHERE cur_user = current_user;
-			INSERT INTO selector_macrosector
-			SELECT DISTINCT macrosector_id, current_user FROM sector WHERE active is true and sector_id IN (SELECT DISTINCT (sector_id) FROM node
-			JOIN selector_expl using (expl_id) where cur_user = current_user)
-			ON CONFLICT (macrosector_id, cur_user) DO NOTHING;
 
 			-- sector
 			DELETE FROM selector_sector WHERE cur_user = current_user;
+
 			INSERT INTO selector_sector
 			SELECT DISTINCT sector_id, current_user FROM node WHERE expl_id IN (SELECT expl_id FROM selector_expl WHERE cur_user = current_user)
 			ON CONFLICT (sector_id, cur_user) DO NOTHING;
@@ -379,12 +373,7 @@ BEGIN
 
 		ELSIF v_tabname IN ('tab_sector', 'tab_macrosector') THEN
 
-			IF v_tabname = 'tab_sector' THEN
-				INSERT INTO selector_macrosector
-				SELECT DISTINCT macrosector_id, current_user FROM sector WHERE sector_id IN (SELECT sector_id FROM selector_sector WHERE cur_user = current_user)
-				ON CONFLICT (macrosector_id, cur_user) DO NOTHING;
-
-			ELSIF v_tabname = 'tab_macrosector' THEN
+			IF  v_tabname = 'tab_macrosector' THEN
 				DELETE FROM selector_sector WHERE cur_user = current_user;
 				INSERT INTO selector_sector
 				SELECT DISTINCT sector_id, current_user FROM sector WHERE macrosector_id IN (SELECT macrosector_id FROM selector_macrosector WHERE cur_user = current_user)
@@ -419,9 +408,6 @@ BEGIN
 
 			-- macroexpl
 			DELETE FROM selector_macroexpl WHERE cur_user = current_user;
-			INSERT INTO selector_macroexpl
-			SELECT DISTINCT macroexpl_id, current_user FROM exploitation WHERE expl_id IN (SELECT DISTINCT expl_id FROM node
-			JOIN selector_municipality using (muni_id) WHERE cur_user = current_user);
 
 			-- expl
 			DELETE FROM selector_expl WHERE cur_user = current_user;
@@ -431,9 +417,6 @@ BEGIN
 
 			-- macrosector
 			DELETE FROM selector_macrosector WHERE cur_user = current_user;
-			INSERT INTO selector_macrosector
-			SELECT DISTINCT macrosector_id, current_user FROM sector WHERE active is true and sector_id IN (SELECT DISTINCT sector_id FROM node
-			JOIN selector_municipality using (muni_id) WHERE cur_user = current_user);
 
 			-- sector
 			DELETE FROM selector_sector WHERE cur_user = current_user AND sector_id > 0;
@@ -482,28 +465,32 @@ BEGIN
 
 		-- macroexpl
 		DELETE FROM selector_macroexpl WHERE cur_user = current_user;
-		INSERT INTO selector_macroexpl
-		SELECT DISTINCT macroexpl_id, current_user FROM exploitation WHERE expl_id IN (SELECT DISTINCT expl_id FROM node
-		JOIN selector_municipality using (muni_id) WHERE cur_user = current_user);
 
 		-- expl
 		DELETE FROM selector_expl WHERE cur_user = current_user;
-		INSERT INTO selector_expl
-		SELECT DISTINCT expl_id, current_user FROM node WHERE muni_id IN (SELECT muni_id FROM selector_municipality WHERE cur_user = current_user)
-		ON CONFLICT (expl_id, cur_user) DO NOTHING;
-
+		IF v_sectorisexplismuni THEN			
+			INSERT INTO selector_expl SELECT expl_id, current_user FROM ud.selector_expl WHERE cur_user = current_user AND expl_id 
+			IN (SELECT expl_id FROM exploitation WHERE active);
+		ELSE	
+			INSERT INTO selector_expl
+			SELECT DISTINCT expl_id, current_user FROM node WHERE muni_id IN (SELECT muni_id FROM selector_municipality WHERE cur_user = current_user)
+			ON CONFLICT (expl_id, cur_user) DO NOTHING;
+		END IF;
+		
 		-- macrosector
 		DELETE FROM selector_macrosector WHERE cur_user = current_user;
-		INSERT INTO selector_macrosector
-		SELECT DISTINCT macrosector_id, current_user FROM sector WHERE active is true and sector_id IN (SELECT DISTINCT sector_id FROM node
-		JOIN selector_municipality using (muni_id) WHERE cur_user = current_user);
 
 		-- sector
 		DELETE FROM selector_sector WHERE cur_user = current_user AND sector_id > 0;
-		INSERT INTO selector_sector
-		SELECT DISTINCT sector_id, current_user FROM node WHERE muni_id IN (SELECT muni_id FROM selector_municipality WHERE cur_user = current_user)
-		ON CONFLICT (sector_id, cur_user) DO NOTHING;
-
+		IF v_sectorisexplismuni THEN			
+			INSERT INTO selector_sector SELECT expl_id, current_user FROM ud.selector_expl WHERE cur_user = current_user AND expl_id 
+			IN (SELECT sector_id FROM sector WHERE active);
+		ELSE 
+			INSERT INTO selector_sector
+			SELECT DISTINCT sector_id, current_user FROM node WHERE muni_id IN (SELECT muni_id FROM selector_municipality WHERE cur_user = current_user)
+			ON CONFLICT (sector_id, cur_user) DO NOTHING;
+		END IF;
+		
 		-- sector for those objects wich has expl_id2 and expl_id2 is not selected but yes one
 		INSERT INTO selector_sector
 		SELECT DISTINCT sector_id,current_user FROM arc WHERE expl_id2 IN (SELECT expl_id FROM selector_expl WHERE cur_user = current_user) AND sector_id > 0
@@ -520,15 +507,6 @@ BEGIN
 		-- psector
 		DELETE FROM selector_psector WHERE psector_id NOT IN
 		(SELECT psector_id FROM cat_dscenario WHERE active is true and expl_id IN (SELECT expl_id FROM selector_expl WHERE cur_user = current_user));
-
-		-- manage explfrommuni
-		IF v_sectorisexplismuni AND v_tabname !='tab_macroexploitation' THEN			
-			DELETE FROM selector_expl WHERE cur_user = current_user;
-			INSERT INTO selector_expl SELECT expl_id, current_user FROM SCHEMA_NAME.selector_expl WHERE cur_user = current_user AND expl_id 
-			IN (SELECT expl_id FROM exploitation WHERE active);
-
-		END IF;		
-		
 	
 		EXECUTE 'SET search_path = '||v_schemaname||', public';
 	END IF;
@@ -556,20 +534,15 @@ BEGIN
 				END IF;
 
 				-- manage value
-				IF v_value then
-					EXECUTE 'INSERT INTO ' || v_tablename || ' ('|| v_columnname ||', cur_user) VALUES('|| v_id ||', '''|| current_user ||''')ON CONFLICT DO NOTHING';
+				IF v_value THEN
+					EXECUTE 'INSERT INTO ' || v_tablename || ' ('|| v_columnname ||', cur_user) VALUES('|| v_id ||', '''|| current_user ||''')ON CONFLICT ('|| v_columnname ||', cur_user) DO NOTHING';
 				ELSE
 					EXECUTE 'DELETE FROM ' || v_tablename || ' WHERE ' || v_columnname || '::text = '''|| v_id ||''' AND cur_user = current_user';
 				END IF;
 
 			END IF;
-
-			IF v_tabname = 'tab_exploitation_add' THEN
-				INSERT INTO selector_macroexpl
-				SELECT DISTINCT macroexpl_id, current_user FROM exploitation WHERE active is true and expl_id IN (SELECT expl_id FROM selector_expl WHERE cur_user = current_user)
-				ON CONFLICT (macroexpl_id, cur_user) DO NOTHING;
-
-			ELSIF v_tabname = 'tab_macroexploitation_add' THEN
+		
+			IF  v_tabname = 'tab_macroexploitation_add' THEN
 				DELETE FROM selector_expl WHERE cur_user = current_user;
 				INSERT INTO selector_expl
 				SELECT DISTINCT expl_id, current_user FROM exploitation WHERE active is true and macroexpl_id IN (SELECT macroexpl_id FROM selector_macroexpl WHERE cur_user = current_user)
@@ -578,16 +551,13 @@ BEGIN
 
 			-- macrosector
 			DELETE FROM selector_macrosector WHERE cur_user = current_user;
-			INSERT INTO selector_macrosector
-			SELECT DISTINCT macrosector_id, current_user FROM sector WHERE active is true and sector_id IN (SELECT DISTINCT (sector_id) FROM node
-			JOIN selector_expl using (expl_id) where cur_user = current_user);
 
 			-- sector
 			DELETE FROM selector_sector WHERE cur_user = current_user AND sector_id > 0;
 			INSERT INTO selector_sector
 			SELECT DISTINCT sector_id, current_user FROM node WHERE expl_id IN (SELECT expl_id FROM selector_expl WHERE cur_user = current_user)
 			ON CONFLICT (sector_id, cur_user) DO NOTHING;
-
+		
 			-- sector for those objects wich has expl_id2 and expl_id2 is not selected but yes one
 			INSERT INTO selector_sector
 			SELECT DISTINCT sector_id,current_user FROM arc WHERE expl_id2 IN (SELECT expl_id FROM selector_expl WHERE cur_user = current_user) AND sector_id > 0
@@ -599,11 +569,13 @@ BEGIN
 			DELETE FROM selector_municipality WHERE cur_user = current_user;
 			INSERT INTO selector_municipality
 			SELECT DISTINCT muni_id, current_user FROM node WHERE expl_id IN (SELECT expl_id FROM selector_expl WHERE cur_user = current_user);
-				
-			EXECUTE' DELETE FROM '||v_schemaname||'.selector_municipality WHERE cur_user = current_user';
-			EXECUTE' INSERT INTO '||v_schemaname||'.selector_municipality 
-			SELECT muni_id, current_user FROM selector_municipality WHERE cur_user = current_user';
-		
+
+			IF v_sectorisexplismuni IS FALSE THEN
+				EXECUTE' DELETE FROM '||v_schemaname||'.selector_municipality WHERE cur_user = current_user';
+				EXECUTE' INSERT INTO '||v_schemaname||'.selector_municipality 
+				SELECT muni_id, current_user FROM selector_municipality WHERE cur_user = current_user';
+			END IF;
+			
 			SELECT row_to_json (a)
 			INTO v_geometry
 			FROM (SELECT st_xmin(the_geom)::numeric(12,2) as x1, st_ymin(the_geom)::numeric(12,2) as y1, 
@@ -617,9 +589,6 @@ BEGIN
 		IF v_sectorisexplismuni IS FALSE THEN
 			-- macroexpl
 			DELETE FROM selector_macroexpl WHERE cur_user = current_user;
-			INSERT INTO selector_macroexpl
-			SELECT DISTINCT macroexpl_id, current_user FROM exploitation WHERE expl_id IN (SELECT DISTINCT expl_id FROM node
-			JOIN selector_municipality using (muni_id) WHERE cur_user = current_user);
 	
 			-- expl
 			DELETE FROM selector_expl WHERE cur_user = current_user;
@@ -629,9 +598,6 @@ BEGIN
 	
 			-- macrosector
 			DELETE FROM selector_macrosector WHERE cur_user = current_user;
-			INSERT INTO selector_macrosector
-			SELECT DISTINCT macrosector_id, current_user FROM sector WHERE active is true and sector_id IN (SELECT DISTINCT sector_id FROM node
-			JOIN selector_municipality using (muni_id) WHERE cur_user = current_user);
 	
 			-- sector
 			DELETE FROM selector_sector WHERE cur_user = current_user AND sector_id > 0;
